remove Sidebar Menu and shadcn code in it + create SidebarNav
