+++ src/components/sidebar/index.ts
@@ -2,4 +2,5 @@ export { default as Sidebar } from './Sidebar.vue';
 export { default as SidebarInset } from './SidebarInset.vue';
 export { default as SidebarProvider } from './SidebarProvider.vue';
 export { default as SidebarTrigger } from './SidebarTrigger.vue';
+export { default as SidebarNav } from './SidebarNav.vue';
